+++ resources/js/screens/monitoring/tag-jobs.vue
@@ -1,4 +1,6 @@
 <script type="text/ecmascript-6">
+    import JobRow from './job-row';
+
     export default {
         props: ['type'],
 
@@ -15,6 +17,13 @@
                 totalPages: 1,
                 jobs: []
             };
+        },
+
+        /**
+         * Components
+         */
+        components: {
+            JobRow,
         },
 
 
@@ -167,43 +176,7 @@
                 </td>
             </tr>
 
-            <tr v-for="job in jobs" :key="job.id">
-                <td>
-                    <span v-if="job.status != 'failed'" :title="job.name">{{jobBaseName(job.name)}}</span>
-                    <router-link v-if="job.status === 'failed'" :title="job.name" :to="{ name: 'failed-jobs-preview', params: { jobId: job.id }}">
-                        {{ jobBaseName(job.name) }}
-                    </router-link><br>
-
-                    <small class="text-muted">
-                        <router-link :to="{name: 'recent-jobs-preview', params: {jobId: job.id}}">View detail</router-link> | 
-                        Queue: {{job.queue}} | Tags: {{ job.payload.tags && job.payload.tags.length ? job.payload.tags.join(', ') : '' }}
-                    </small>
-                </td>
-                <td class="table-fit">
-                    {{ readableTimestamp(job.payload.pushedAt) }}
-                </td>
-
-                <td class="table-fit" v-if="type == 'jobs'">
-                    <span>{{ job.completed_at ? (job.completed_at - job.reserved_at).toFixed(2)+'s' : '-' }}</span>
-                </td>
-
-                <td class="text-right table-fit" v-if="type == 'jobs'">
-                    <svg v-if="job.status == 'completed'" class="fill-success" viewBox="0 0 20 20" style="width: 1.5rem; height: 1.5rem;">
-                        <path d="M2.93 17.07A10 10 0 1 1 17.07 2.93 10 10 0 0 1 2.93 17.07zm12.73-1.41A8 8 0 1 0 4.34 4.34a8 8 0 0 0 11.32 11.32zM6.7 9.29L9 11.6l4.3-4.3 1.4 1.42L9 14.4l-3.7-3.7 1.4-1.42z"></path>
-                    </svg>
-
-                    <svg v-if="job.status == 'reserved' || job.status == 'pending'" class="fill-warning" viewBox="0 0 20 20" style="width: 1.5rem; height: 1.5rem;">
-                        <path d="M2.93 17.07A10 10 0 1 1 17.07 2.93 10 10 0 0 1 2.93 17.07zm12.73-1.41A8 8 0 1 0 4.34 4.34a8 8 0 0 0 11.32 11.32zM7 6h2v8H7V6zm4 0h2v8h-2V6z"/>
-                    </svg>
-
-                    <svg v-if="job.status == 'failed'" class="fill-danger" viewBox="0 0 20 20" style="width: 1.5rem; height: 1.5rem;">
-                        <path d="M2.93 17.07A10 10 0 1 1 17.07 2.93 10 10 0 0 1 2.93 17.07zm1.41-1.41A8 8 0 1 0 15.66 4.34 8 8 0 0 0 4.34 15.66zm9.9-8.49L11.41 10l2.83 2.83-1.41 1.41L10 11.41l-2.83 2.83-1.41-1.41L8.59 10 5.76 7.17l1.41-1.41L10 8.59l2.83-2.83 1.41 1.41z"/>
-                    </svg>
-                </td>
-
-                <td class="text-right table-fit" v-if="type == 'failed'">
-                    {{ readableTimestamp(job.failed_at) }}
-                </td>
+            <tr v-for="job in jobs" :key="job.id" :job="job" is="job-row">
             </tr>
             </tbody>
         </table>